requirements.txt
ffmpeg-python
numpy
opencv-python
scenedetect[opencv]
insightface
onnxruntime
torch
torchvision
librosa
tqdm
PyYAML
